+++ include/GafferScene/RendererAlgo.h
@@ -73,8 +73,10 @@ GAFFERSCENE_API void createOutputDirectories( const IECore::CompoundObject *glob
 GAFFERSCENE_API void transformSamples( const ScenePlug *scene, size_t segments, const Imath::V2f &shutter, std::vector<Imath::M44f> &samples, std::vector<float> &sampleTimes );
 
 /// Samples the object from the current location in preparation for output to the renderer. Sampling parameters
-/// are as for the transformSamples() method. Multiple samples will only be generated for Primitives, since other
-/// object types cannot be interpolated anyway.
+/// are as for the transformSamples() method. Multiple samples will only be generated for Primitives and Cameras,
+/// since other object types cannot be interpolated anyway.
+GAFFERSCENE_API void objectSamples( const ScenePlug *scene, size_t segments, const Imath::V2f &shutter, std::vector<IECore::ConstObjectPtr> &samples, std::vector<float> &sampleTimes );
+/// \deprecated
 GAFFERSCENE_API void objectSamples( const ScenePlug *scene, size_t segments, const Imath::V2f &shutter, std::vector<IECoreScene::ConstVisibleRenderablePtr> &samples, std::vector<float> &sampleTimes );
 
 /// Function to return a SceneProcessor used to adapt the